requirements.txt
# Web (Streamlit)
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Data Collection
apify-client>=1.0.0  # Apify for Reddit
google-search-results>=2.4.2  # SerpAPI
openai>=1.0.0  # OpenAI API for embeddings and LLM

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Worker (스케줄링 등)
# schedule>=1.2.0
# celery>=5.3.0  # 선택사항

# Machine Learning
hdbscan>=0.8.33  # HDBSCAN clustering
scikit-learn>=1.3.0  # TF-IDF, utilities
